% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{as.serialdate}
\alias{as.serialdate}
\title{Convert a date to an integer to match Excel}
\usage{
as.serialdate(x)
}
\arguments{
\item{x}{a date object, or anything that can be cast as a numeric}
}
\value{
Returns a vector of values with the same length as \code{x}.
Where Jan 1, 1970, returns 25569 to match Excel's internal
representation of dates.
}
\description{
Convert a date to an integer to match Excel
}
\examples{
as.serialdate(as.Date('2025-02-28'))
as.serialdate(as.Date('2025-03-01'))
as.serialdate('1970-01-01')
as.serialdate('1900-03-01')
as.serialdate('1900-01-01')
as.serialdate(as.Date('1970-01-01'):as.Date('1970-12-31')) - 25569

}
