% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{flatten_whitespace}
\alias{flatten_whitespace}
\title{Flatten whitespace}
\usage{
flatten_whitespace(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
Returns a vector of values with the same length as \code{x}.
All whitespace is replaced with single
}
\description{
Flatten any whitespace to be copy-paste safe, meaning no
tab or newline characters
}
\examples{
flatten_whitespace(' This is a newline\nAnd now it is gone! ')
flatten_whitespace(' This is a newline\n\rAnd now it is gone! ')
flatten_whitespace(' This is a newline\r\n\n\nAnd now it is gone! ')
flatten_whitespace(' This is a newline\n\n\nAnd now it is gone! ')

}
