% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{info}
\alias{info}
\title{Tabular summary of data.frame columns}
\usage{
info(x, ...)
}
\arguments{
\item{dataframe}{anything that can be coerced into a dataframe}
}
\value{
Like the summary function, but tabular and with more information
}
\description{
Create a tabular form of summary info
}
\examples{
info(iris)

}
