% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{intersection}
\alias{intersection}
\title{Intersection of sets}
\usage{
intersection(...)
}
\arguments{
\item{...}{as many vectors as you want}
}
\value{
A vector of values with no duplicates.
Each value in the return vector appeared in all input vectors.
}
\description{
Give me a bunch of vectors.
I'll give you a unique list of all the values
that appear in all of them.
}
\examples{
v1 <- 1:5
v2 <- 2:10
v3 <- c("1", "2", "3")
intersection(v1, v2)
intersection(v2, v3)
intersection(v1, v2, v3)

}
