% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{left}
\alias{left}
\title{Get a substring from the left}
\usage{
left(string, n)
}
\arguments{
\item{string}{a vector of strings}

\item{n}{a vector of integers (will be recycled, if needed)}
}
\value{
The first \code{n} characters of each value in \code{string}.
If \code{n} has length 1, then that value is applied to all strings
because of recycling.
However, \code{n} can have more than one element!
}
\description{
Inspired by the Excel function with the same name.
}
\examples{
left('lazy dog', 4)
left(c('lazy dog', 'blue dog'), 4)
left(c('lazy dog', 'red dog'), c(4, 3))

}
