% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{load_env}
\alias{load_env}
\title{Load Rdata to an environment}
\usage{
load_env(RData, env = new.env())
}
\arguments{
\item{RData}{path to the RData file you want to load}

\item{env}{the environment you want to load to.  Defaults to a new environment.}
}
\value{
Returns the environment you are loading to.
}
\description{
Instead of loading RData files to the global environment
load them to a new env.
\code{base::load} already has a way to do this,
but this is a convenience function that creates
a new environment by default.
}
\examples{
# data_env <- load_env('file1.RData')
# load_env('file2.RData', data_env)

}
