% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{matrix2latex}
\alias{matrix2latex}
\title{Doc writing: get LaTeX of a matrix}
\usage{
matrix2latex(M, one_line = TRUE, print = TRUE, ret = !print)
}
\arguments{
\item{M}{a matrix-like object}

\item{one_line}{logical -- whether to return one-line or many}

\item{print}{logical -- print the LaTeX to the screen}

\item{ret}{logical -- return the text so it can be stored in a variable}
}
\value{
If \code{ret = TRUE}, the LaTeX code to represent the matrix.
If \code{print = TRUE}, the code is printed to the screen using \code{message}.
}
\description{
If you are working in RMarkdown and need to write a matrix in LaTeX
it can be a pain.  This function tries to make it easier.
}
\examples{
m <- datasets::iris[1:5, 1:4]
matrix2latex(m)
tex <- matrix2latex(m, print = FALSE)
print(tex)

}
