% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{mid}
\alias{mid}
\title{Get a substring from the middle}
\usage{
mid(string, start, length)
}
\arguments{
\item{string}{a vector of strings
start     a vector of starting points (will be recycled, if needed)
length    a vector of lengths (will be recycled, if needed)}
}
\value{
The \code{length} characters of each value in \code{string}
starting with the \code{start} character.
}
\description{
Inspired by the Excel function with the same name.
}
\examples{
mid('lazy dog', 3, 2)
mid('lazy dog', 6, 3)
mid(c('lazy dog', 'blue dog'), 2, 2)
mid(c('lazy dog', 'red frog'), 6, c(3, 2))

}
