% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{month}
\alias{month}
\title{Month of a date vector}
\usage{
month(d)
}
\arguments{
\item{d}{a vector of dates}
}
\value{
The month part of each value in \code{d}
as an integer.
}
\description{
Month of a date vector
}
\examples{
month(as.Date('2020-10-14'))
month(ISOdate(1923, 12, 2))

}
