% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{normalize_quote_characters}
\alias{normalize_quote_characters}
\title{Replace all quote-like characters with straight quotes}
\usage{
normalize_quote_characters(x)
}
\value{
Returns a character vector with quote-likes replaced.
}
\description{
Use `gsub` to replace all quote-like characters,
including smart quotes and non-English quote characters,
with straight English quotes: `"` or `'`.
See https://stackoverflow.com/a/47173868.
}
\examples{
x <- c(
  '\u00AB',  # \u00AB  LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
  '\u00BB',  # \u00BB  RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
  '\u05F4',  # \u05F4  HEBREW PUNCTUATION GERSHAYIM
  '\u201C',  # \u201C  LEFT DOUBLE QUOTATION MARK
  '\u201D',  # \u201D  RIGHT DOUBLE QUOTATION MARK
  '\u201E',  # \u201E  DOUBLE LOW-9 QUOTATION MARK
  '\u201F',  # \u201F  DOUBLE HIGH-REVERSED-9 QUOTATION MARK
  '\u226A',  # \u226A  MUCH LESS-THAN
  '\u226B',  # \u226B  MUCH GREATER-THAN
  '\u300A',  # \u300A  LEFT DOUBLE ANGLE BRACKET
  '\u300B',  # \u300B  RIGHT DOUBLE ANGLE BRACKET
  '\u301D',  # \u301D  REVERSED DOUBLE PRIME QUOTATION MARK
  '\u301E',  # \u301E  DOUBLE PRIME QUOTATION MARK
  '\u301F',  # \u301F  LOW DOUBLE PRIME QUOTATION MARK
  '\uFF02',  # \uFF02  FULLWIDTH QUOTATION MARK
  '\u2033',  # \u2033  DOUBLE PRIME
  '\u2036',  # \u2036  REVERSED DOUBLE PRIME
  '\u02BB',  # \u02BB  MODIFIER LETTER TURNED COMMA
  '\u02BC',  # \u02BC  MODIFIER LETTER APOSTROPHE
  '\u02BD',  # \u02BD  MODIFIER LETTER REVERSED COMMA
  '\u066C',  # \u066C  ARABIC THOUSANDS SEPARATOR
  '\u2018',  # \u2018  LEFT SINGLE QUOTATION MARK
  '\u2019',  # \u2019  RIGHT SINGLE QUOTATION MARK
  '\u201A',  # \u201A  SINGLE LOW-9 QUOTATION MARK
  '\u201B',  # \u201B  SINGLE HIGH-REVERSED-9 QUOTATION MARK
  '\u055A',  # \u055A  ARMENIAN APOSTROPHE
  '\uFE10',  # \uFE10  PRESENTATION FORM FOR VERTICAL COMMA
  '"',
  "'")

y <- data.frame(
  all_quotes = x,
  straight_quotes = normalize_quote_characters(x))

print(y)
table(y$straight_quotes)

}
