% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{open_test}
\alias{open_test}
\title{Package dev: open a test file}
\usage{
open_test(test, pkg = basename(getwd()), engine = "testthat")
}
\arguments{
\item{test}{the object you want to write a test for}

\item{pkg}{to pass for the knitr environment.  Defaults to current working directory}

\item{engine}{if you are using something other than testthat,
you can save this test to "tests/ENGINE/"}
}
\value{
Returns NULL invisibly.
Side-effect: opens the test file in the editor.
}
\description{
For use in package development.
Open a file in tests/testthat/ for the object called "test"
If the file doesn't already exist, create it.
If it already exists, open it for editing.
}
\examples{
# open_test(lm)

}
