% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{pmean}
\alias{pmean}
\title{Piecewise mean}
\usage{
pmean(..., na.rm = TRUE)
}
\arguments{
\item{...}{vectors}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\value{
Return the average of the numbers in the same position
across all inputs.
}
\description{
\code{pmean} is to \code{\link[base]{mean}} what
\code{\link[base]{pmax}} is to \code{\link[base]{max}}.
}
\examples{
pmean(1:10, 5, 11:20)
pmean(1:10, 5, c(11:19, NA))
pmean(1:10, 5, c(11:19, NA), na.rm = FALSE)

}
