% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{quarter}
\alias{quarter}
\title{Quarter of a date vector}
\usage{
quarter(d)
}
\arguments{
\item{d}{a vector of dates}
}
\value{
The quarter (1, 2, 3, or 4) each date in \code{d}
falls in.
}
\description{
Quarter of a date vector
}
\examples{
quarter(as.Date('2020-01-14'))
quarter(ISOdate(1923, 12, 2))

}
