% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{quarter_name}
\alias{quarter_name}
\title{Quarter name of a date vector, e.g., 2023q2}
\usage{
quarter_name(d)
}
\arguments{
\item{d}{a vector of dates}
}
\value{
Year and quarter concatenated delimited by the letter 'q'.
}
\description{
Quarter name of a date vector, e.g., 2023q2
}
\examples{
quarter_name(as.Date('2022-01-14'))
quarter_name(ISOdate(1923, 12, 2))

}
