% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{read.csv0}
\alias{read.csv0}
\title{Read CSVs without factors}
\usage{
read.csv0(
  file,
  header = TRUE,
  sep = ",",
  quote = "\\"",
  dec = ".",
  fill = TRUE,
  comment.char = "",
  ...,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{...}{See parameters for \code{\link{read.csv}}}
}
\value{
See help page for \code{\link{read.csv}}
}
\description{
\code{read.csv0} is to \code{read.csv} what
\code{paste0} is to \code{paste}.
More often that not, I see people using \code{read.csv}
with \code{stringsAsFactors = FALSE} (like you will
often see people calling \code{paste} with \code{sep = ''}).
This is a variant where \code{stringsAsFactors} is preset.
Relies on the \code{\link{curry}} function.
}
\examples{
# See help page for \code{\link{read.csv}}

}
