% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{repair_names}
\alias{repair_names}
\title{Fix column names to be like conventional relational database names}
\usage{
repair_names(x)
}
\arguments{
\item{x}{names to convert}
}
\value{
Returns fixed names guaranteed to be unique
}
\description{
Fix column names to be like conventional relational database names
}
\examples{
x <- c(
  "DR_NO", "Quarter", "Date Rptd",
  "DATE OCC", "TIME OCC", "AREA", "AREA NAME",
  "Rpt Dist No", "Part 1-2", "Crm Cd", "Crm Cd Desc", "Mocodes",
  "Vict Age", "Vict Sex", "Vict Descent", "Premis Cd", "Premis Desc",
  "Weapon Used Cd", "Weapon Desc", "Status", "Status Desc", "Crm Cd 1",
  "Crm Cd 2", "Crm Cd 3", "Crm Cd 4", "LOCATION|().,{}''", "Cross Street",
  "LAT", "LON", "FXEXCHANGERATE...COM",
  "Currency Symbols     ", "Currency Symbols - FX Exchange Rate",
  "Home#", "Currency___Converter!",
  "World__Countries", "Gold\%", '    ', "Currency Codes(ISO)",
  "100", "m$", "£", "€", "QUARTER")
repair_names(x)

}
