% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{right}
\alias{right}
\title{Get a substring from the right}
\usage{
right(string, n)
}
\arguments{
\item{string}{a vector of strings}

\item{n}{a vector of integers (will be recycled, if needed)}
}
\value{
The last \code{n} characters of each value in \code{string}.
If \code{n} has length 1, then that value is applied to all strings
because of recycling.
However, \code{n} can have more than one element!
}
\description{
Inspired by the Excel function with the same name.
}
\examples{
right('lazy dog', 3)
right(c('lazy dog', 'blue dog'), 3)
right(c('lazy dog', 'red frog'), c(3, 4))

}
