% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{source_env}
\alias{source_env}
\title{Source R code to an environment}
\usage{
source_env(RScript, env = new.env())
}
\arguments{
\item{RScript}{path to the R file you want to run}

\item{env}{the environment you want to run in.  Defaults to a new environment.}
}
\value{
Returns the environment where the processing occurred.
}
\description{
Instead of sourcing R code in the global environment
where it might create new objects or overwrite existing ones,
create a new environment (or use an existing one)
and source the code there!
}
\examples{
# code_env <- source_env('file1.R')
# source_env('file2.R', code_env)

}
