% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{table0}
\alias{table0}
\title{Table function where `useNA` is defaulted to `TRUE`}
\usage{
table0(...)
}
\arguments{
\item{...}{arguments passed through to the `table` function}
}
\value{
Returns output of `table` but forces `NA` to always be counted
}
\description{
Identical to `table(..., useNA = "always")`.
The name is inspired by `paste0`.
}
\examples{
x <- sample(c(letters, NA), 1000, replace = TRUE)
table(x)
table(x, useNA = "always")
table0(x)

}
