% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{vlookup}
\alias{vlookup}
\title{Excel-like \code{VLOOKUP} function}
\usage{
vlookup(
  lookup_value,
  table_array,
  col_index_number,
  type = 0,
  lookup_index = 1
)
}
\arguments{
\item{lookup_value}{a vector of values you want to match in the \code{table_array} column of interest}

\item{table_array}{the data you are "looking up" in}

\item{col_index_number}{the index of the column you want to return}

\item{type}{0 = exact match.  No other method is currently supported}

\item{lookup_index}{the column index of the column \code{lookup_value} is matched to
(In Excel, this is always "1")}
}
\value{
Returns a vector of values with the same length as \code{lookup_value}.
If a value is not found, then NA is returned.
Unlike Excel, the column we are matching against cannot have duplicates.
}
\description{
Works just like Excel's \code{VLOOKUP} function
with a few improvements.
}
\examples{
ref <- data.frame(
  state = c('UT', 'FL', 'NY', 'CA', 'SD'),
  category = c(1, 2, 3, 3, 1)
)

vlookup('CA', ref, 1, 0)
vlookup('CA', ref, 2, 0)

vlookup(c('CA', 'FL', 'KY'), ref, 2, 0)

ref[2, 2] <- 4
vlookup('FL', ref, 2, 0)


}
