% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{write.function}
\alias{write.function}
\title{Write the code of a function to a file}
\usage{
write.function(f, file, name = NULL)
}
\arguments{
\item{TODO}{Note}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}
