% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{write.package}
\alias{write.package}
\title{Write the code of a package to a file}
\usage{
write.package(pkg, folder)
}
\arguments{
\item{pkg}{the package of interest}

\item{folder}{the folder where to save the R files}
}
\value{
Nothing
}
\description{
Uses \code{\link{write.package}} to write
every function in a package to a specified directory.
}
\examples{
# write.package('alrtools', 'some/folder/')

}
