% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{xlookup}
\alias{xlookup}
\title{Excel-like \code{XLOOKUP} function}
\usage{
xlookup(
  x,
  lookup_vector,
  return_vector = NULL,
  warn = TRUE,
  ignore_case = TRUE
)
}
\arguments{
\item{x}{a vector of values you want to match in \code{lookup_vector}}

\item{lookup_vector}{the vector to match in -- get the index for \code{return_vector}}

\item{return_vector}{the vector to return from}

\item{warn}{TRUE will list values of \code{x} that are not found in \code{lookup_vector}}

\item{ignore_case}{TRUE will do a case insensitive match of \code{x} on \code{lookup_vector}}
}
\value{
Returns a vector of values with the same length as \code{lookup_vector}.
If a value is not found, then NA is returned.
NA can be mapped to a non-NA value.
}
\description{
Works like Excel's \code{XLOOKUP} function
}
\examples{
ref <- data.frame(
  state = c('UT', 'SD', 'FL', NA, 'NY', 'CA'),
  category = c(1, 2, 5, 10, 6, 7)
)

xlookup('CA', ref$state, ref$category)
xlookup('KY', ref$state, ref$category)
xlookup(NA, ref$state, ref$category)

}
