% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{year}
\alias{year}
\title{Year of a date vector}
\usage{
year(d)
}
\arguments{
\item{d}{a vector of dates}
}
\value{
The year part of each value in \code{d}
as an integer.
}
\description{
Year of a date vector
}
\examples{
year(as.Date('2020-10-14'))
year(ISOdate(1923, 12, 2))

}
