% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{ifelse}
\alias{ifelse}
\title{A factor-safe version of \code{\link[base]{ifelse}}}
\usage{
ifelse(test, yes, no)
}
\arguments{
\item{test}{a vector of logicals
yes    values to return if test is TRUE
no     values to return if test if FALSE}
}
\value{
For each position, \code{ifelse} returns
the value of \code{yes} or \code{no}
depending on the value of \code{test}.
}
\description{
The \code{\link[base]{ifelse}} function does weird
things when you pass it factors.
This version doesn't.
}
\examples{
ifelse(
  c(TRUE, FALSE),
  c('Yes', 'Yeah!'),
  c('No', 'Nope'))

# base ifelse does weird stuff with factors
yesf <- factor(c('Yes', 'Yeah!'))
nof <- factor(c('No', 'Nope'))
base::ifelse(c(T, F), yesf, nof)
ifelse(c(T, F), yesf, nof)

}
