% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{load_csv}
\alias{load_csv}
\title{Load CSV with info about original header}
\usage{
load_csv(
  path,
  col_types = NULL,
  na = c("", "NA", "#N/A", "#DIV/0!", "#NAME?", "#REF!", "#VALUE!", "#NULL!"),
  ...
)
}
\arguments{
\item{path}{path to the CSV file to load}

\item{col_types}{One of NULL, a cols() specification, or a string. See vignette("readr") for more details.}

\item{na}{values to replace with NA -- defaults to Excel errors}

\item{...}{other arguments passed to readr::read_csv}
}
\value{
Returns a tibble of the CSV file
}
\description{
A wrapper for readr::read_csv that uses alrtools::repair_header
to fix the header, but keeps a copy of the original values in an attribute.
}
\examples{
my_data <- data.frame(
  'Seq No' = 1:26,
  'Uppercase Letters' = LETTERS,
  'Lowercase Letters' = letters,
  check.names = FALSE)

temp_path <- tempfile(fileext = '.csv')
write.csv(x = my_data, file = temp_path, row.names = FALSE)

my_data_reloaded <- load_csv(temp_path)

# Fixed Names
names(my_data_reloaded)

# Original Names
attr(my_data_reloaded, 'original_header')

# Clean Up
unlink(temp_path)

}
