% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{matrix2latex}
\alias{matrix2latex}
\title{Doc writing: get LaTeX of a matrix}
\usage{
matrix2latex(M, one_line = TRUE, print = TRUE, ret = !print)
}
\arguments{
\item{TODO}{Note}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}
