% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{open_test}
\alias{open_test}
\title{Package dev: open a test file}
\usage{
open_test(test, pkg = basename(getwd()), engine = "testthat")
}
\arguments{
\item{TODO}{Note}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO

}
